+++ data_structures/union_find.hpp
@@ -5,18 +5,18 @@
 
 class union_find {
     private:
-    size_t n;
+    int n;
     std::vector<int> p;
 
     public:
     union_find() : n(0) {}
-    explicit union_find(size_t _n) : n(_n), p(n, -1) {}
+    explicit union_find(int _n) : n(_n), p(n, -1) {}
 
-    unsigned int root(unsigned int a) { return p[a] < 0 ? a : (p[a] = root(p[a])); }
+    int root(int a) { return p[a] < 0 ? a : (p[a] = root(p[a])); }
 
-    bool connected(unsigned int a, unsigned int b) { return root(a) == root(b); }
+    bool connected(int a, int b) { return root(a) == root(b); }
 
-    unsigned int connect(unsigned int a, unsigned int b) {
+    int connect(int a, int b) {
         a = root(a); b = root(b);
         if (a == b) return a;
         if (p[a] > p[a]) std::swap(a, b);
@@ -24,9 +24,9 @@ class union_find {
         return a;
     }
 
-    size_t size() const { return n; }
+    int size() const { return n; }
 
-    size_t size(unsigned int a) { return -p[root(a)]; }
+    int size(int a) { return -p[root(a)]; }
 };
 
 #endif // DATA_STRUCTURES_UNION_FIND_HPP
